refactor: extract runtime outcome policy
